Added Math Functions
With these, rookies and other users of Gesture.cs can do more
complicated gestures such as simply saying if something is within a
certain 2D space, or a 3D one. If a programmer asks if a Vector is a
certain 3D distance from another Vector, the user can move the limb
freely within that space and still get a value of true.

>2D Distance
>3D Distance
>Slope